requirements.txt
langchain
langchain-openai
pandas
openpyxl
python-dotenv